+++ src/sqlfluff/core/parser/__init__.py
@@ -17,7 +17,6 @@ from .grammar import (
     Sequence,
     GreedyUntil,
     StartsWith,
-    ContainsOnly,
     OneOf,
     Delimited,
     Bracketed,